Move to :core:data test.
