+++ Compose-Flight-Search-App/core/data/src/test/kotlin/com/example/flightsearchapp/core/data/repository/SearchTextRepositoryTest.kt
@@ -1,12 +1,11 @@
-package com.example.flightsearchapp.repository
+package com.example.flightsearchapp.core.data.repository
 
-import com.example.flightsearchapp.data.SearchTextRepository
-import com.example.flightsearchapp.fake.SearchTextFakeRepository
+import com.example.flightsearchapp.core.testing.fake.datasource.SearchTextFakeRepository
+import kotlinx.coroutines.flow.first
+import kotlinx.coroutines.test.runTest
 import kotlin.test.BeforeTest
 import kotlin.test.Test
 import kotlin.test.assertEquals
-import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.test.runTest
 
 /**
  * Instrumented test for [InDiskSearchTextRepository][SearchTextRepository]
